Implement accuracy as stopping criteria
Previously we used a pre-set number of iterations through the data set for learning the weights in the perceptron algorithm. This was replaced by using accuracy as the stopping/convergence criteria where if the current set of weights results in an accuracy greater than or equal to a pre-set threshold we stop training.